refactor(97020): Extrai código remoção acessos para área comum
Para ser reutilizado também no formulário de usuários, o código
referente ao tratamento de remoção de acessos foi extraído para uma
posta comum às páginas der lista e form de usuários.
